implementation of casts 1: hookup to valueSwitchStmt
